+++ src/main/java/com/global/api/terminals/upa/Entities/Enums/UpaSearchCriteria.java
@@ -3,5 +3,10 @@ package com.global.api.terminals.upa.Entities.Enums;
 public enum UpaSearchCriteria {
     EcrId,
     Batch,
-    ReportOutput
+    ReportOutput,
+    ReportType,
+    ReportSubType,
+    BothReports,
+    ClerkId,
+    PreviousBatchReport
 }